+++ rust/fishsense-core/src/laser/calibration.rs
@@ -8,8 +8,8 @@ pub enum LaserCalibrationError {
     CentroidCalculationError,
 }
 
-pub fn calibrate_laser(points: &Array2<f32>) -> Result<(ndarray::Array1<f32>, ndarray::Array1<f32>), LaserCalibrationError> {
-    let mut laser_orientation = Array1::<f32>::zeros(3);
+pub fn calibrate_laser(points: &Array2<f64>) -> Result<(ndarray::Array1<f64>, ndarray::Array1<f64>), LaserCalibrationError> {
+    let mut laser_orientation = Array1::<f64>::zeros(3);
     let &[n, _] = points.shape() else {
         panic!("Shape is not 2-dimensional!") // Panic is okay here since the dimensions are known at compile time
     };
